hawk: change folder structure and call exec_monitor from the C++ program
