+++ src/include/exec_monitor.h
@@ -1,3 +1,6 @@
+#ifndef EXEC_MONITOR_H
+#define EXEC_MONITOR_H
+
 #include <linux/limits.h>
 
 struct process_info {
@@ -6,3 +9,7 @@ struct process_info {
 	int tgid;
 	char name[PATH_MAX];
 };
+
+void exec_monitor();
+
+#endif
