+++ src/kernelspace/exec_monitor.c
@@ -17,7 +17,7 @@
 #include <bpf/bpf_helpers.h>
 #include <bpf/bpf_tracing.h>
 #include <bpf/bpf_core_read.h>
-#include "process_info.h"
+#include "exec_monitor.h"
 
 struct {
 	__uint(type, BPF_MAP_TYPE_RINGBUF);
